feat: refactor the code
